document-management-system-main/src/main/java/com/sample/app/service/UserService.java
package com.sample.app.service;

import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

import com.sample.app.dto.UserDto;
import com.sample.app.entity.User;
import com.sample.app.repository.UserRepository;

@Service
public class UserService {
	private final UserRepository userRepository;

	@Autowired
	public UserService(UserRepository userRepository) {
		this.userRepository = userRepository;
	}

	public User createUser(UserDto userDto) {
		User user = new User();
		user.setName(userDto.getName());
		return userRepository.save(user);
	}

	public User getUserById(Long userId) {
		return userRepository.findById(userId)
				.orElseThrow(() -> new RuntimeException("User not found with ID: " + userId));
	}

}
